+++ linuxdoom-1.10/ib_video/xlib.c
@@ -465,7 +465,7 @@ static void I_Quit_Wrapper(int dummy)
     I_Quit();
 }
 
-void IB_InitGraphics(const char *title, size_t screen_width, size_t screen_height, unsigned int *bytes_per_pixel)
+void IB_InitGraphics(const char *title, size_t screen_width, size_t screen_height, size_t *bytes_per_pixel)
 {
 
     char*		displayname;
